# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TraderLoading API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/check-name:
    get:
      operationId: checkProfileName
      tags: [profile]
      summary: Check if a profile name is available
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Name availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckNameResponse"
  /profile/avatar:
    post:
      operationId: uploadProfileAvatar
      tags: [profile]
      summary: Upload a profile avatar image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        "200":
          description: Uploaded avatar URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
  /profile/avatar/generate:
    post:
      operationId: generateProfileAvatar
      tags: [profile]
      summary: Generate a unique AI avatar
      responses:
        "200":
          description: Generated avatar URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
        "500":
          description: Generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        xpToNextLevel:
          type: integer
        streak:
          type: integer
        levelName:
          type: string
        yearsExperience:
          type: integer
          nullable: true
        winRate:
          type: number
          nullable: true
          description: "Win rate percentage computed from journal entries (0-100)"
        totalTrades:
          type: integer
          nullable: true
      required:
        - id
        - name
        - xp
        - level
        - xpToNextLevel
        - streak
        - levelName
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        yearsExperience:
          type: integer
          nullable: true
      required:
        - name
    CheckNameResponse:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available
    AvatarResponse:
      type: object
      properties:
        avatarUrl:
          type: string
      required:
        - avatarUrl
...